Added client options and connected to server
Added client options. There's a bug where the ClientHandler doesn't actually send back the response, it might be in the Server threadPool code; im not sure how that command works. Also, the client inputs are not currently being checked/implemented; for example, the friends list isn't created. Last note: the database file needs to be written to with the information so that the data persists through multiple runs.